chore: prepare PrivateKey for new key types
